refactor(providers): split responses adapter into per-direction files
internal/providers/responses_adapter.go was 1005 lines covering four
distinct directions of Responses↔Chat translation. Split into peer
files in the same package:

  responses_adapter.go (179)  ChatProvider iface, request->chat,
                              tool/tool-choice normalization,
                              ResponsesViaChat / StreamResponsesViaChat
  responses_input.go (297)    Input items -> Chat messages,
                              assistant-message merging,
                              stringify/firstNonEmpty helpers
  responses_content.go (367)  Content normalization, image/audio
                              part conversion, ExtractContentFromInput,
                              extras-extraction helpers
  responses_output.go (185)   Output items + ConvertChatResponseToResponses,
                              ResponsesFunctionCall{CallID,ItemID}

Pure relocation; tests pass.
